hugging_faces/requirements.txt
transformers>=4.37.0
torch>=2.0.0
accelerate>=0.25.0
sentencepiece>=0.1.99
requests>=2.28.0
python-dotenv>=1.0.0
streamlit>=1.28.0
Pillow>=9.0.0
streamlit-mic-recorder>=0.0.8
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
pydantic>=2.0.0
chromadb>=0.4.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.2.0
mcp>=1.20.0
